requirements.txt
azure-ai-inference==1.0.0b2
openai==1.35.14
mistralai==0.4.2
python-dotenv==1.0.1